+++ custom_components/soundtouchplus/const.py
@@ -3,14 +3,16 @@ Constants for the Bose SoundTouchPlus component.
 """
 DOMAIN = "soundtouchplus"
 
-CONF_WS_PORT = "ws_port"
 CONF_DEVICE_NAME = "device_name"
 CONF_DEVICE_ID = "device_id"
+CONF_PING_WEBSOCKET_INTERVAL = "ping_websocket_interval"
+CONF_PORT_WEBSOCKET = "port_websocket"
 
+DEFAULT_PING_WEBSOCKET_INTERVAL = 300
 DEFAULT_PORT = 8090
+DEFAULT_PORT_WEBSOCKET = 8080
 DEFAULT_SSL = False
 DEFAULT_TIMEOUT = 15
-DEFAULT_WS_PORT = 8080
 
 # custom service names.
 SERVICE_PLAY_HANDOFF = "play_handoff"